Dor e sofrimento
CREATE DATABASE escola_db;
use escola_db;
CREATE TABLE curso (
    curso_id BIGINT AUTO_INCREMENT PRIMARY KEY,
    nome_curso VARCHAR(100) NOT NULL,
    descricao TEXT,
    carga_horaria INT NOT NULL
);

CREATE TABLE componente (
    componente_id BIGINT AUTO_INCREMENT PRIMARY KEY,
    nome_componente VARCHAR(100) NOT NULL,
    carga_horaria INT NOT NULL,
    curso_id BIGINT NOT NULL,
    CONSTRAINT fk_componente_curso
	    FOREIGN KEY (curso_id)
	    REFERENCES curso(curso_id)
	    ON DELETE CASCADE
	    ON UPDATE CASCADE
);

CREATE TABLE turma (
    turma_id BIGINT AUTO_INCREMENT PRIMARY KEY,
    componente_id BIGINT NOT NULL,
    ano INT NOT NULL,
    semestre TINYINT NOT NULL,
    turno VARCHAR(20) NOT NULL,
    capacidade INT NOT NULL,
    CONSTRAINT fk_turma_componente
	    FOREIGN KEY (componente_id)
	    REFERENCES componente(componente_id)
	    ON DELETE CASCADE
        ON UPDATE CASCADE
);

CREATE TABLE aluno (
    aluno_id BIGINT AUTO_INCREMENT PRIMARY KEY,
    nome_aluno VARCHAR(150) NOT NULL,
    cpf VARCHAR(14) UNIQUE,
    data_nascimento DATE NOT NULL,
    email VARCHAR(100),
    telefone VARCHAR(20) NOT NULL
);

CREATE TABLE matricula (
    matricula_id BIGINT AUTO_INCREMENT PRIMARY KEY,
    aluno_id BIGINT NOT NULL,
    turma_id BIGINT NOT NULL,
    data_matricula DATE DEFAULT (CURRENT_DATE),
    CONSTRAINT fk_matricula_aluno
        FOREIGN KEY (aluno_id)
        REFERENCES aluno(aluno_id)
        ON DELETE CASCADE,
    CONSTRAINT fk_matricula_turma
        FOREIGN KEY (turma_id)
        REFERENCES turma(turma_id)
        ON DELETE CASCADE,
    CONSTRAINT unq_aluno_turma UNIQUE (aluno_id, turma_id)
);

+++ bispo/src/main/java/br/com/bispo/model/CursoModel.java
@@ -23,7 +23,7 @@ public class CursoModel {
 	private String descricao;
 	
 	@Column(name = "carga_horaria")
-	private Long cargaHoraria;
+	private Integer cargaHoraria;
 
 	public Long getCursoId() {
 		return cursoId;
@@ -49,11 +49,11 @@ public class CursoModel {
 		this.descricao = descricao;
 	}
 
-	public Long getCargaHoraria() {
+	public Integer getCargaHoraria() {
 		return cargaHoraria;
 	}
 
-	public void setCargaHoraria(Long cargaHoraria) {
+	public void setCargaHoraria(Integer cargaHoraria) {
 		this.cargaHoraria = cargaHoraria;
 	}
